pyproject.toml
[tool.poetry]
name = "commenthub-sqr"
version = "0.1.0"
description = ""
authors = ["dmitriypru <dmitriypru@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.2"
uvicorn = {extras = ["standart"], version = "^0.29.0"}
sqlalchemy = "^2.0.29"
bcrypt = "^4.1.2"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
streamlit = "^1.33.0"
beautifulsoup4 = "^4.12.3"
pydantic = "^2.7.1"
email-validator = "^2.1.1"
python-dotenv = "^1.0.1"
python-multipart = "^0.0.9"


[tool.poetry.group.test.dependencies]
pytest = "^8.2.0"
coverage = "^7.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
